refactor: make markdown component
- styledMarkdown.tsxに統合
