--- web/app/components/markdown/blockquote.tsx
@@ -1,9 +1,0 @@
-import type React from "react";
-
-const Blockquote: React.FC<
-	React.BlockquoteHTMLAttributes<HTMLQuoteElement>
-> = ({ children, ...props }) => {
-	return <blockquote {...props}>{children}</blockquote>;
-};
-
-export default Blockquote;
