--- web/app/components/markdown/index.ts
@@ -1,11 +1,0 @@
-export { default as A } from "./a";
-export { default as Blockquote } from "./blockquote";
-export { default as Code } from "./code";
-export { default as H1 } from "./h1";
-export { default as H2 } from "./h2";
-export { default as H3 } from "./h3";
-export { default as Hr } from "./hr";
-export { default as Li } from "./li";
-export { default as Ol } from "./ol";
-export { default as P } from "./p";
-export { default as Ul } from "./ul";
