--- web/app/components/markdown/ol.tsx
@@ -1,10 +1,0 @@
-import type React from "react";
-
-const Ol: React.FC<React.OlHTMLAttributes<HTMLOListElement>> = ({
-	children,
-	...props
-}) => {
-	return <ol {...props}>{children}</ol>;
-};
-
-export default Ol;
